Add permission checks to controllers

+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 use Inertia\Inertia;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
@@ -11,6 +13,10 @@ class RoleController extends Controller
 {
     public function index()
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         return Inertia::render('Admin/Roles/Index', [
             'roles' => Role::with('permissions')->get(),
         ]);
@@ -18,6 +24,10 @@ class RoleController extends Controller
 
     public function create()
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         return Inertia::render('Admin/Roles/Create', [
             'permissions' => Permission::all(),
         ]);
@@ -25,6 +35,10 @@ class RoleController extends Controller
 
     public function store(Request $request)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $data = $request->validate([
             'name' => 'required|unique:roles,name',
             'permissions' => 'array',
@@ -38,6 +52,10 @@ class RoleController extends Controller
 
     public function edit(Role $role)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         return Inertia::render('Admin/Roles/Edit', [
             'role' => $role->load('permissions'),
             'permissions' => Permission::all(),
@@ -46,6 +64,10 @@ class RoleController extends Controller
 
     public function update(Request $request, Role $role)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $data = $request->validate([
             'name' => 'required|unique:roles,name,' . $role->id,
             'permissions' => 'array',
@@ -59,6 +81,10 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $role->delete();
         return redirect()->route('roles.index')->with('success', 'Papel excluído com sucesso.');
     }
